Add ChipInput component
TC-48 is a rework of how institution emails are entered. Currently, they are comma separated and there is no real validation that this is correctly entered by the users.

To improve UX with minimal changes under the hood I've introduced a new type of input called ChipInput that supports multiple elements and returns an array under the hood (that we can then serialize however we like).

The original concept was to wrap an open source library called 'material-ui-chip-input'. However this proved to have problems due to the way our components are styled. Instead I've reimplemented most of the relevant functionality from 'material-ui-chip-input' and added styling, sorting and validation.

After discussions with Zoe, validation is positive (light blue) and unvalidated entries are simply grey.