Allow sharing of class name generator across each ThemeProvider
In TC we render multiple react trees, because the React must live
alongside legacy code.

This means that in TC we need to be able to share the class name
generator across each instance of a ThemeProvider. If the generator
is not shared, it's highly likely that the generated classnames will
clash and cause components to have the incorrect styles applied to
them.

Adding this prop allows us to share the same instance of name
generator and avoid this problem.