Allow MessageTileButton component to accept formatting props
A plan of the MessageTile is to be more flexible corss usage.

To alow the MessageTile to fully replace the soon-to-be-defunkt
`ArticleDonationBanner` component, we had to allow two extra props.

`fullWidth` is a material UI prop, making the button run full width.
We use this for the republishing example, but for the end of article
call-to-action we shouldn't have the button run full width as the component
is displayed wider than some other examples.

`prominent` is one of our concepts, and it gives the `Button` extra padding
to the left and the right of it's label text.

I've set up defaults matching expected usage