Add a StickyPopup component, which build a Snackbar under the hood
We want to allow for less-intrusive popups to alert readers to
donation campaigns or newsletters. Unlike larger dialog-based
popups, the `StickyPopup` doesn't add a 'curtain' or drop-shadow
covering the rest of the page.

It's "sticky" because it, like a `Snackbar` is bound to a corner of
the page.

It accepts `children` components for full customisability.