Add clickAwayClose prop, disabling clickaway behaviour be default.
The StickyPopup (more specifically, the underlying Snackbar) by-default
had behaviour that clicking away closed the popup.

This is because it wasn't _meant_ to have a dismissable "X".

So, now, by default the click away handlers do nothing.
We can also optionally set them to when clicked away, it closes the StickyPopup.