Add a ThemeSelector component
The `<ThemeSelector>` is a bridge between a string (the name of a theme),
and the theme object itself.

It's job is to know about the themes, alleviating consuming apps, (like TC),
from having to know about the themes.

I think this is especially useful in the case of Promos, where we're storing the theme
as a string.

To be used in place of a ThemeProvider, as it wraps a ThemeProvider under the hood.