Add resets to heading typography variants
This isn't ideal, but it's the easiest way to prevent TC's legacy
CSS from messing with MUI typography heading variants.

Initially I planned to reset every property that we set in TC, but
that caused additional problems due to class specificity and the
way MUI mixes classes together to build the styles for Typography
components.

In the end, I experimented until I only had resets for just the
properties that were likely to cause (and have caused) the most
grief while effectively being a no-op visually.