Remove now-unused ThinDonationBanner component
We had donation-focused hardcoded (in that the colours were hardcoded red, rather than using themes) when we launched promos to get moving quickly

Later, we built themeable components to replace them.
In this components case, ThinBanner.

We've migrated TC from using the hardcoded banners, and need to remove them from the UI library.