Pass id to Select in Dropdown
With the upgrade to version 2.41.0 of this library, testing library in
donations has effectively [started
complaining](https://buildkite.com/conversation/tc-donations/builds/14481#7fcee22d-9468-4f7d-a293-f8e987799f72)
that TC-UI `Select` elements do not have  ids (it can't find the
associated form control for the label the spec is searching for, because
the associated form control doesn't have the id we're passing in).

Given a `Dropdown` component with an id of `"xyz"`, the DOM should
include a `select` with an id of `"xyz"` as well as a label with a `for`
attribute of `"xyz"`.